Added Side bar to settings panel
