+++ src/components/panels/settings/sections/GeneralSection.tsx
@@ -1,0 +1,292 @@
+import { useEffect, useState } from "react";
+import { invoke } from "@tauri-apps/api/core";
+import { openUrl } from "@tauri-apps/plugin-opener";
+import type { AppInfo } from "../../../../store";
+import { changelogEntries } from "../../../../changelog";
+import ChangelogContent from "../../../ChangelogContent";
+import ConfirmDialog from "../../../ConfirmDialog";
+import { SettingsCard } from "./shared";
+
+interface CumulativeStats {
+  totalClicks: number;
+  totalTimeSecs: number;
+  totalSessions: number;
+  avgCpu: number;
+}
+
+function formatTime(totalSeconds: number, language: string): string {
+  if (totalSeconds < 0.01) return "0s";
+  if (totalSeconds < 60) {
+    return `${Math.floor(totalSeconds).toLocaleString(language)}s`;
+  }
+  if (totalSeconds < 3600) {
+    const m = Math.floor(totalSeconds / 60);
+    const s = Math.floor(totalSeconds % 60);
+    return s > 0
+      ? `${m.toLocaleString(language)}m ${s.toLocaleString(language)}s`
+      : `${m.toLocaleString(language)}m`;
+  }
+  const h = Math.floor(totalSeconds / 3600);
+  const m = Math.floor((totalSeconds % 3600) / 60);
+  return m > 0
+    ? `${h.toLocaleString(language)}h ${m.toLocaleString(language)}m`
+    : `${h.toLocaleString(language)}h`;
+}
+
+function formatNumber(n: number, language: string): string {
+  return Math.floor(n).toLocaleString(language);
+}
+
+function formatCpu(
+  cpu: number,
+  language: string,
+  notAvailable: string,
+): string {
+  if (cpu < 0) return notAvailable;
+  return `${cpu.toLocaleString(language, {
+    minimumFractionDigits: 1,
+    maximumFractionDigits: 1,
+  })}%`;
+}
+
+interface Props {
+  appInfo: AppInfo;
+  updateCheckStatus:
+    | "idle"
+    | "checking"
+    | "available"
+    | "unavailable"
+    | "error";
+  onCheckForUpdate: () => void;
+}
+
+export default function GeneralSection({
+  appInfo,
+  updateCheckStatus,
+  onCheckForUpdate,
+}: Props) {
+  const [stats, setStats] = useState<CumulativeStats | null>(null);
+  const [showChangelog, setShowChangelog] = useState(false);
+  const [pendingAction, setPendingAction] = useState<"clear-stats" | null>(
+    null,
+  );
+  const [resettingStats, setResettingStats] = useState(false);
+  const language = "en";
+
+  useEffect(() => {
+    invoke<CumulativeStats>("get_stats")
+      .then(setStats)
+      .catch(() => {});
+  }, []);
+
+  const hasStats = stats !== null && stats.totalSessions > 0;
+
+  const handleConfirmClearStats = async () => {
+    setResettingStats(true);
+    try {
+      const next = await invoke<CumulativeStats>("reset_stats");
+      setStats(next);
+    } catch {
+      // stats unchanged on failure
+    } finally {
+      setResettingStats(false);
+      setPendingAction(null);
+    }
+  };
+
+  const updateButtonLabel = {
+    idle: "Check for Update",
+    checking: "Checking...",
+    available: "Update found!",
+    unavailable: "No update found",
+    error: "Check failed",
+  }[updateCheckStatus];
+
+  return (
+    <>
+      <SettingsCard title="About" description="Version and project links.">
+        <div className="social-links">
+          <span className="settings-label">Support Me</span>
+          <div className="social-icons">
+            <a
+              className="social-icon social-icon--kofi"
+              href="#"
+              title="Ko-fi"
+              onClick={(e) => {
+                e.preventDefault();
+                void openUrl("https://ko-fi.com/Z8Z71T8QD4");
+              }}
+            >
+              <img
+                height="28"
+                style={{ border: 0, height: "28px" }}
+                src="https://storage.ko-fi.com/cdn/kofi3.png?v=6"
+                alt="Buy Me a Coffee at ko-fi.com"
+              />
+            </a>
+            <a
+              className="social-icon social-icon--youtube"
+              href="#"
+              title="YouTube"
+              onClick={(e) => {
+                e.preventDefault();
+                void openUrl("https://youtube.com/@Blur009");
+              }}
+            >
+              <svg
+                viewBox="0 0 24 24"
+                fill="currentColor"
+                width="18"
+                height="18"
+              >
+                <path d="M23.498 6.186a3.016 3.016 0 0 0-2.122-2.136C19.505 3.545 12 3.545 12 3.545s-7.505 0-9.377.505A3.017 3.017 0 0 0 .502 6.186C0 8.07 0 12 0 12s0 3.93.502 5.814a3.016 3.016 0 0 0 2.122 2.136c1.871.505 9.376.505 9.376.505s7.505 0 9.377-.505a3.015 3.015 0 0 0 2.122-2.136C24 15.93 24 12 24 12s0-3.93-.502-5.814zM9.545 15.568V8.432L15.818 12l-6.273 3.568z" />
+              </svg>
+            </a>
+            <a
+              className="social-icon social-icon--twitch"
+              href="#"
+              title="Twitch"
+              onClick={(e) => {
+                e.preventDefault();
+                void openUrl("https://twitch.tv/Blur009");
+              }}
+            >
+              <svg
+                viewBox="0 0 24 24"
+                fill="currentColor"
+                width="18"
+                height="18"
+              >
+                <path d="M11.571 4.714h1.715v5.143H11.57zm4.715 0H18v5.143h-1.714zM6 0L1.714 4.286v15.428h5.143V24l4.286-4.286h3.428L22.286 12V0zm14.571 11.143l-3.428 3.428h-3.429l-3 3v-3H6.857V1.714h13.714z" />
+              </svg>
+            </a>
+            <a
+              className="social-icon social-icon--github"
+              href="#"
+              title="GitHub"
+              onClick={(e) => {
+                e.preventDefault();
+                void openUrl("https://github.com/Blur009/Blur-AutoClicker");
+              }}
+            >
+              <svg
+                viewBox="0 0 24 24"
+                fill="currentColor"
+                width="18"
+                height="18"
+              >
+                <path d="M12 .3a12 12 0 0 0-3.8 23.4c.6.1.8-.2.8-.6v-2c-3.3.7-4-1.4-4-1.4-.5-1.3-1.2-1.7-1.2-1.7-1-.7.1-.7.1-.7 1.1.1 1.7 1.2 1.7 1.2 1 .1.8 1.8 3.4 1.2.1-.7.4-1.2.7-1.5-2.7-.3-5.4-1.3-5.4-6a4.7 4.7 0 0 1 1.2-3.2c-.1-.3-.5-1.5.1-3.2 0 0 1-.3 3.3 1.2a11.2 11.2 0 0 1 6.1 0c2.3-1.5 3.3-1.2 3.3-1.2.6 1.7.2 2.9.1 3.2a4.7 4.7 0 0 1 1.2 3.2c0 4.7-2.8 5.7-5.4 6 .4.3.8 1 .8 2.1v3.1c0 .4.2.7.8.6A12 12 0 0 0 12 .3" />
+              </svg>
+            </a>
+          </div>
+        </div>
+        <div className="settings-row">
+          <div className="settings-label-group settings-label-group--inline">
+            <span className="settings-label">Version</span>
+            <span className="settings-value">v{appInfo.version}</span>
+          </div>
+          <div className="settings-row-actions">
+            <button
+              className="settings-btn-secondary changelog-toggle-btn"
+              onClick={() => setShowChangelog((v) => !v)}
+            >
+              <svg
+                className={`changelog-arrow${showChangelog ? " open" : ""}`}
+                width="10"
+                height="10"
+                viewBox="0 0 10 10"
+                fill="none"
+              >
+                <path
+                  d="M3 1L7 5L3 9"
+                  stroke="currentColor"
+                  strokeWidth="1.5"
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                />
+              </svg>
+              {showChangelog ? "Hide Changes" : "Show Changes"}
+            </button>
+            <button
+              className="settings-btn-secondary check-update-btn"
+              onClick={onCheckForUpdate}
+              disabled={updateCheckStatus !== "idle"}
+            >
+              {updateButtonLabel}
+            </button>
+          </div>
+        </div>
+        {showChangelog && <ChangelogContent entries={changelogEntries} />}
+      </SettingsCard>
+
+      <SettingsCard
+        title="Usage"
+        description="Clicking statistics for all sessions."
+      >
+        <div className="settings-row">
+          <div className="settings-label-group">
+            <span className="settings-label">Usage Data</span>
+            <span className="settings-sublabel">
+              Statistics are stored locally and never sent anywhere.
+            </span>
+          </div>
+        </div>
+        {hasStats ? (
+          <div className="stats-grid">
+            <div className="stats-cell">
+              <span className="stats-cell-label">Total Clicks</span>
+              <span className="stats-cell-value">
+                {formatNumber(stats.totalClicks, language)}
+              </span>
+            </div>
+            <div className="stats-cell">
+              <span className="stats-cell-label">Total Time Clicking</span>
+              <span className="stats-cell-value">
+                {formatTime(stats.totalTimeSecs, language)}
+              </span>
+            </div>
+            <div className="stats-cell">
+              <span className="stats-cell-label">Average CPU</span>
+              <span className="stats-cell-value">
+                {formatCpu(stats.avgCpu, language, "N/A")}
+              </span>
+            </div>
+            <div className="stats-cell">
+              <span className="stats-cell-label">Sessions</span>
+              <span className="stats-cell-value">
+                {formatNumber(stats.totalSessions, language)}
+              </span>
+            </div>
+          </div>
+        ) : (
+          <div className="stats-empty">No session data yet.</div>
+        )}
+        {hasStats && (
+          <div className="settings-row">
+            <div className="settings-label-group">
+              <span className="settings-label">Clear Stats</span>
+              <span className="settings-sublabel">Clear all usage data.</span>
+            </div>
+            <button
+              type="button"
+              className="settings-btn-danger settings-btn-danger--compact"
+              onClick={() => setPendingAction("clear-stats")}
+            >
+              Clear
+            </button>
+          </div>
+        )}
+      </SettingsCard>
+
+      <ConfirmDialog
+        open={pendingAction === "clear-stats"}
+        title="Clear usage data"
+        message="This will clear all usage data. This action cannot be undone."
+        confirmLabel="Clear"
+        busy={resettingStats}
+        onConfirm={handleConfirmClearStats}
+        onCancel={() => setPendingAction(null)}
+      />
+    </>
+  );
+}
